Add files via upload
MUSEFUL is a cross-plattform MERNstack application to learn, search for, create and collect content – in a musefully and playful way. In the one hand this CRUD grounded fullstack project gets for example the vocabulary learning component words from APIs like Datamuse, Wordnik or Oxford Dictionary. On the other hand this application could be used to create and store user-generated list.